feat: run benchmarks with double types by default
The scalapack benchmarks use doubles, so to avoid
future mistakes and enhance comparability, I decided
to use doubles here as well.

+++ benchmarks/buffer_benchmarks.cpp
@@ -25,7 +25,7 @@
 
 #include <reshuffle.hpp>
 
-using SendType = int;
+using SendType = double;
 
 double time_shuffle(const std::vector<SendType> &values,
                     const reshuffle::BlockCyclic &current_distribution,
